refactor: moved everything into utilities folder and deleted constants folder
